+++ lib/startback/context.rb
@@ -11,6 +11,13 @@ module Startback
   class Context
     attr_accessor :original_rack_env
 
+    # An error handler can be provided on the Context class. The latter
+    # MUST expose an API similar to ruby's Logger class. It can be a logger
+    # instance, simply.
+    #
+    # Fatal errors catched by Web::CatchAll are sent on `error_handler#fatal`
+    attr_accessor :error_handler
+
   end # class Context
 end # module Startback
 require_relative 'context/middleware'