+++ lib/startback/web/catch_all.rb
@@ -8,6 +8,9 @@ module Startback
     # This class aims at being used as top level of a Rack chain. It is not
     # aimed at being subclassed.
     #
+    # Fatal error cached are also sent as a `fatal` messange, on the error
+    # handler provided on Context#error_handler.fatal, if any.
+    #
     # Examples:
     #
     #     Rack::Builder.new do
@@ -18,17 +21,23 @@ module Startback
       include Errors
 
       FATAL_ERROR = {
-        code: "Gybr::Errors::ServerError",
+        code: "Startback::Errors::InternalServerError",
         description: "An error occured, sorry"
       }.to_json
 
       self.catch_all
+      self.on(Exception)
       self.status 500
       self.content_type 'application/json'
-      self.body FATAL_ERROR.to_json
+      self.body FATAL_ERROR
 
       self.ensure(true) do |ex|
-        Startback::LOGGER.fatal(ex)
+        context = env[Context::Middleware::RACK_ENV_KEY]
+        if context && context.respond_to?(:error_handler) && context.error_handler
+          context.error_handler.fatal(ex)
+        else
+          Startback::LOGGER.fatal(ex)
+        end
       end
 
     end # class CatchAll